refactor response size limiter
