+++ pkg/querier/tripperware/queryrange/query_range.go
@@ -23,6 +23,7 @@ import (
 	"github.com/cortexproject/cortex/pkg/querier/tripperware"
 	"github.com/cortexproject/cortex/pkg/util"
 
+	"github.com/cortexproject/cortex/pkg/util/limiter"
 	"github.com/cortexproject/cortex/pkg/util/spanlogger"
 )
 
@@ -206,7 +207,7 @@ func (c prometheusCodec) DecodeResponse(ctx context.Context, r *http.Response, _
 		return nil, err
 	}
 
-	responseSizeLimiter := tripperware.ResponseSizeLimiterFromContextWithFallback(ctx)
+	responseSizeLimiter := limiter.ResponseSizeLimiterFromContextWithFallback(ctx)
 
 	if strings.EqualFold(r.Header.Get("Content-Encoding"), "gzip") && len(buf.Bytes()) >= 4 {
 		// Read the uncompressed gzip response size from the footer
